Moved all filter Gui Packets to base
TODO move existing item handler to use only IItemHandler
