+++ enderio-base/src/main/java/crazypants/enderio/base/filter/gui/SpeciesItemFilterGui.java
@@ -11,7 +11,9 @@ import com.enderio.core.client.gui.button.ToggleButton;
 import crazypants.enderio.base.EnderIO;
 import crazypants.enderio.base.filter.filters.SpeciesItemFilter;
 import crazypants.enderio.base.filter.filters.SpeciesMode;
+import crazypants.enderio.base.filter.network.PacketFilterUpdate;
 import crazypants.enderio.base.gui.IconEIO;
+import crazypants.enderio.base.network.PacketHandler;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.tileentity.TileEntity;
@@ -117,7 +119,8 @@ public class SpeciesItemFilterGui extends AbstractGuiItemFilter {
 
   private void sendFilterChange() {
     updateButtons();
-    filterContainer.onFilterChanged();
+    PacketHandler.INSTANCE
+        .sendToServer(new PacketFilterUpdate(filterContainer.getTileEntity(), filter, filterContainer.filterIndex, filterContainer.getParam1()));
   }
 
   @Override
