+++ enderio-base/src/main/java/crazypants/enderio/base/network/PacketHandler.java
@@ -8,7 +8,7 @@ import crazypants.enderio.base.EnderIO;
 import crazypants.enderio.base.config.PacketConfigSync;
 import crazypants.enderio.base.config.PacketConfigSyncNew;
 import crazypants.enderio.base.config.PacketConfigSyncNew.PacketConfigSyncNewHandler;
-import crazypants.enderio.base.filter.gui.PacketFilterUpdate;
+import crazypants.enderio.base.filter.network.PacketFilterUpdate;
 import crazypants.enderio.base.handler.darksteel.PacketDarkSteelPowerPacket;
 import crazypants.enderio.base.handler.darksteel.PacketDarkSteelSFXPacket;
 import crazypants.enderio.base.handler.darksteel.PacketUpgradeState;
