+++ enderio-conduits/src/main/java/crazypants/enderio/conduit/item/ItemConduit.java
@@ -31,7 +31,6 @@ import crazypants.enderio.base.render.registry.TextureRegistry;
 import crazypants.enderio.base.render.registry.TextureRegistry.TextureSupplier;
 import crazypants.enderio.base.tool.ToolUtil;
 import crazypants.enderio.conduit.AbstractConduit;
-import crazypants.enderio.conduit.AbstractConduitNetwork;
 import crazypants.enderio.conduit.IConduitComponent;
 import crazypants.enderio.conduit.gui.GuiExternalConnection;
 import crazypants.enderio.conduit.gui.item.ItemSettings;
@@ -557,7 +556,7 @@ public class ItemConduit extends AbstractConduit implements IItemConduit, ICondu
   }
 
   @Override
-  public @Nullable AbstractConduitNetwork<?, ?> getNetwork() {
+  public @Nullable ItemConduitNetwork getNetwork() {
     return network;
   }
