+++ enderio-conduits/src/main/java/crazypants/enderio/conduit/packet/PacketHandler.java
@@ -4,6 +4,8 @@ import javax.annotation.Nonnull;
 
 import com.enderio.core.common.network.ThreadedNetworkWrapper;
 
+import crazypants.enderio.base.filter.network.PacketExistingItemFilterSnapshot;
+import crazypants.enderio.base.filter.network.PacketModItemFilter;
 import crazypants.enderio.conduit.EnderIOConduits;
 import crazypants.enderio.conduit.liquid.PacketConduitFluidLevel;
 import net.minecraft.entity.player.EntityPlayerMP;
